+++ alicompat/CHANGELOG.md
@@ -1,3 +1,7 @@
+## []
+
+- Added GregTech CEu Modern support
+
 ## [1.0.0]
 
 - Initial release